app/image-resizer-appv1.0/app/zip-processing/zip-processor/requirements.txt
confluent-kafka
pillow
requests
python-json-logger
